refactor: session관련 폴더 위치 수정 (#32)
