+++ src/main/java/pokemon/pokedex/user/service/UserServiceImpl.java
@@ -3,8 +3,8 @@ package pokemon.pokedex.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import pokemon.pokedex._global.session.SessionConst;
-import pokemon.pokedex._global.session.registry.SessionRegistry;
+import pokemon.pokedex._global.SessionConst;
+import pokemon.pokedex._common.session.registry.SessionRegistry;
 import pokemon.pokedex.user.domain.AdminRequestStatus;
 import pokemon.pokedex.user.dto.SessionUserDTO;
 import pokemon.pokedex.user.repository.UserRepository;
