+++ src/test/java/pokemon/pokedex/_common/session/SessionListenerTest.java
@@ -1,4 +1,4 @@
-package pokemon.pokedex._global.session;
+package pokemon.pokedex._common.session;
 
 import jakarta.servlet.http.HttpSession;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,7 +11,7 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import pokemon.pokedex.ClearMemory;
-import pokemon.pokedex._global.session.registry.SessionRegistry;
+import pokemon.pokedex._common.session.registry.SessionRegistry;
 import pokemon.pokedex.user.dto.LoginDTO;
 import pokemon.pokedex.user.dto.RegisterDTO;
 import pokemon.pokedex.user.dto.SessionUserDTO;
