+++ src/test/java/pokemon/pokedex/user/controller/LoginControllerUnitTest.java
@@ -14,8 +14,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import pokemon.pokedex.WebMvcTestWithExclude;
-import pokemon.pokedex._global.session.SessionConst;
-import pokemon.pokedex._global.session.registry.SessionRegistry;
+import pokemon.pokedex._global.SessionConst;
+import pokemon.pokedex._common.session.registry.SessionRegistry;
 import pokemon.pokedex.user.dto.LoginDTO;
 import pokemon.pokedex.user.dto.RegisterResponseDTO;
 import pokemon.pokedex.user.dto.SessionUserDTO;
