+++ src/test/java/pokemon/pokedex/user/interceptor/GuestOnlyInterceptorUnitTest.java
@@ -14,8 +14,8 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import pokemon.pokedex.WebMvcTestWithExclude;
-import pokemon.pokedex._global.session.SessionConst;
-import pokemon.pokedex._global.session.registry.SessionRegistry;
+import pokemon.pokedex._global.SessionConst;
+import pokemon.pokedex._common.session.registry.SessionRegistry;
 import pokemon.pokedex.user.controller.LoginController;
 import pokemon.pokedex.user.controller.RegisterController;
 import pokemon.pokedex.user.dto.LoginDTO;
