+++ src/test/java/pokemon/pokedex/user/interceptor/LoginUserInjectInterceptorUnitTest.java
@@ -10,7 +10,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import pokemon.pokedex.WebMvcTestWithExclude;
 import pokemon.pokedex._common.HomeController;
-import pokemon.pokedex._global.session.SessionConst;
+import pokemon.pokedex._global.SessionConst;
 import pokemon.pokedex.user.domain.AdminRequestStatus;
 import pokemon.pokedex.user.domain.Role;
 import pokemon.pokedex.user.dto.SessionUserDTO;
