+++ src/test/java/pokemon/pokedex/user/service/UserServiceImplTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import pokemon.pokedex._global.session.SessionConst;
-import pokemon.pokedex._global.session.registry.SessionRegistry;
+import pokemon.pokedex._global.SessionConst;
+import pokemon.pokedex._common.session.registry.SessionRegistry;
 import pokemon.pokedex.user.domain.AdminRequestStatus;
 import pokemon.pokedex.user.domain.User;
 import pokemon.pokedex.user.dto.SessionUserDTO;
